refactor: Convert codebase to a class-based structure
This commit refactors the entire codebase to use a more organized,
object-oriented structure.

- Creates a new `scripts/classes` directory to house the new classes.
- Introduces `ExperimentReader`, `DataProcessor`, `GrowthAnalyzer`, and
  `Plotter` classes to encapsulate the different parts of the workflow.
- Moves the relevant functions from the old `scripts/functions` directory
  and `scripts/growth.py` into these new classes as methods.
- Refactors the main `scripts/growth.py` script to be a simple
  orchestrator of the new class objects.
- Removes the old `scripts/functions` directory.
