+++ scripts/classes/reader.py
@@ -1,4 +1,3 @@
-# readers.py
 import pandas as pd
 
 class ExperimentReader:
