+++ scripts/growth.py
@@ -17,68 +17,21 @@ __email__ = 'dmartimarti **AT** gmail.com'
 __maintainer__ = 'Daniel Martínez Martínez'
 __status__ = 'beta'
 __date__ = 'Dec 2024'
-__version__ = '0.8.0'
+__version__ = '0.9.0'
 
-from functions.functions import *
-from functions.readers import ExperimentReader
-from functions.analyzers import GrowthAnalyzer
+from classes.reader import ExperimentReader
+from classes.processor import DataProcessor
+from classes.analyzer import GrowthAnalyzer
+from classes.plotter import Plotter
 import argparse
 import pathlib
-import numpy as np
-from scipy import signal
-from tqdm import tqdm
+import os
 import shutil
 from itertools import product
 from multiprocessing import get_context
-import sys
+from tqdm import tqdm
+import pandas as pd
 
-from scipy import interpolate as ip
-from sklearn.decomposition import PCA
-
-# start parser
-parser = argparse.ArgumentParser()
-
-# define arguments to pass to the script
-# here comes one difference, I'll read a path instead of a Design file
-parser.add_argument('-i',
-                    '--input',
-                    type=lambda p: pathlib.Path(p).absolute(),
-                    help='Input path where the design file is located (in .xlsx format)',
-                    required=True)
-parser.add_argument('-f',
-                    '--file',
-                    default='Design.xlsx',
-                    help='Name of the input file. "Design.xlsx" by default.')
-parser.add_argument('-o',
-                    '--output',
-                    default='Output',
-                    help='Output folder to save the analysis')
-parser.add_argument('-t',
-                    '--threads',
-                    default=1,
-                    help='Number of threads to use. 1 by default.')
-# add an argument to specify if you are using windows, optional
-parser.add_argument('-w',
-                    '--windows',
-                    action='store_true',
-                    help='Flag to indicate that you are using Windows.')
-parser.add_argument('-v',
-                    '--version',
-                    action='version',
-                    version=f'%(prog)s v.{__version__} {__status__}')
-parser.add_argument('--parametric',
-                    action='store_true',
-                    help='Flag to enable parametric modelling.')
-parser.add_argument('--pca',
-                    action='store_true',
-                    help='Flag to enable PCA analysis.')
-parser.add_argument('--pca_color',
-                    type=str,
-                    default=None,
-                    help='Variable to color the PCA plot by.')
-
-
-args = parser.parse_args()
 
 # class of colors to print in the terminal
 class bcolors:
@@ -93,96 +46,55 @@ class bcolors:
     UNDERLINE = '\033[4m'
 
 
-class AnalysisVars():
-    def __init__(self, analysis_vars):
-        self.grouping_variable = analysis_vars.grouping_variable.dropna().to_list()
-        self.condition = analysis_vars.condition.dropna().to_list()[0] 
-        self.all_vars = self.grouping_variable + [self.condition]
-    
-    # method to append any ammount of elements to the list in grouping_variable
-    def append_grouping_variable(self, *args):
-        temp_list = self.grouping_variable
-        for arg in args:
-            temp_list.append(arg)
-        return temp_list
-
-
-# initialise variables
-ROOT = args.input
-OUTPUT = args.output
-n_threads = int(args.threads)
-
-
-reader = ExperimentReader(os.path.join(ROOT, args.file))
-design = reader.read_design()
-
-analyzer = GrowthAnalyzer(design, ROOT)  # Pass necessary arguments
-
-# this function calculates the AUC of the OD time series
-def out_auc_function(file, mode):
-    # open the info about df, temperatures, times and OD
-    df, temps, OD = analyzer.biospa_text_opener(os.path.join(ROOT, file))
-    OD = OD[3:]
-
-    if analyzer.check_outliers_temps(temps):
-        raise Exception("There are outliers in the temperatures, check your experiment!")
-    else:
-        pass
-
-    # get time info
-    length, time_h, time_span = analyzer.get_time_h(df)
-
-    ### Fix and interpolate the data
-    window = int(round(length / 10, 0))
-    # make all the time series start from the same point (0 usually)
-    adj_df = df.apply(lambda row: pd.Series(analyzer.set_series2min(row - np.mean(row[:window]), 0.0), index=time_span), axis=1)
-
-    # smooth the data from adj_df using a wiener filter
-    with np.errstate(divide='ignore', invalid='ignore'): # ignore the warnings from wiener filter
-        w_filt = adj_df.apply(lambda row: pd.Series(signal.wiener(row, 5), index=time_span), axis=1)
-    # fix the NaN rows and substitute them with 0
-    w_filt = w_filt.fillna(0)
-
-    growth_rates = w_filt.apply(lambda row: pd.Series(ip.UnivariateSpline(time_h, row, s=0).derivative(1)(time_h), index=time_span),
-                    axis=1)
-
-    max_slope = growth_rates.apply(lambda row: pd.Series(row.rolling(window).mean().max(), index=time_span), axis=1).iloc[:,0]
-
-    # calculate AUC from the smoothed data
-    auc = w_filt.apply(lambda row: pd.Series(np.trapz(row, time_h), index=time_span), axis=1).iloc[:,0]
-
-    ######## prepare the Output.csv
-    # remove RuntimeWarning: divide by zero encountered in log2
-    with np.errstate(divide='ignore', invalid='ignore'):
-        auc_log2 = np.log2(auc)
-
-    auc_df = pd.DataFrame({f'File': file, f'{OD}_f_AUC': auc, f'{OD}_f_logAUC': auc_log2, f'{OD}_dt_Max': max_slope})
-    auc_df.reset_index(inplace=True)
-    auc_df = auc_df.rename(columns = {0:'Well'})
-
-    # get the timeseries
-    # add metadata to adj_df
-    adj_df.insert(0, 'File', file)
-    adj_df.insert(1, 'Data', f'{OD}nm')
-    adj_df.insert(2, 'Well', adj_df.index)
-    # add metadata to w_filt
-    w_filt.insert(0, 'File', file)
-    w_filt.insert(1, 'Data', f'{OD}nm_f')
-    w_filt.insert(2, 'Well', w_filt.index)
-    # add metadata to growth_rates
-    growth_rates.insert(0, 'File', file)
-    growth_rates.insert(1, 'Data', f'{OD}nm_dt')
-    growth_rates.insert(2, 'Well', growth_rates.index)
-
-    # concat the dfs
-    timeseries_df = pd.concat([adj_df, w_filt, growth_rates], axis=0)
-
-    if mode == 'AUC':
-        return auc_df
-    elif mode == 'timeseries':
-        return timeseries_df
-
 def main():
+    parser = argparse.ArgumentParser()
+
+    # define arguments to pass to the script
+    # here comes one difference, I'll read a path instead of a Design file
+    parser.add_argument('-i',
+                        '--input',
+                        type=lambda p: pathlib.Path(p).absolute(),
+                        help='Input path where the design file is located (in .xlsx format)',
+                        required=True)
+    parser.add_argument('-f',
+                        '--file',
+                        default='Design.xlsx',
+                        help='Name of the input file. "Design.xlsx" by default.')
+    parser.add_argument('-o',
+                        '--output',
+                        default='Output',
+                        help='Output folder to save the analysis')
+    parser.add_argument('-t',
+                        '--threads',
+                        default=1,
+                        help='Number of threads to use. 1 by default.')
+    # add an argument to specify if you are using windows, optional
+    parser.add_argument('-w',
+                        '--windows',
+                        action='store_true',
+                        help='Flag to indicate that you are using Windows.')
+    parser.add_argument('-v',
+                        '--version',
+                        action='version',
+                        version=f'%(prog)s v.{__version__} {__status__}')
+    parser.add_argument('--parametric',
+                        action='store_true',
+                        help='Flag to enable parametric modelling.')
+    parser.add_argument('--pca',
+                        action='store_true',
+                        help='Flag to enable PCA analysis.')
+    parser.add_argument('--pca_color',
+                        type=str,
+                        default=None,
+                        help='Variable to color the PCA plot by.')
+
+    args = parser.parse_args()
+
+    # initialise variables
+    ROOT = args.input
+    OUTPUT = args.output
+    n_threads = int(args.threads)
+
     print(f'The folder {os.path.split(args.input)[-1]} will be analysed.')
 
     # if -w is passed, store a variable with the string "spawn"
@@ -191,17 +103,25 @@ def main():
     else:
         start_method = 'fork'
 
+    # Initialize classes
+    reader = ExperimentReader(os.path.join(ROOT, args.file))
+    processor = DataProcessor()
+    analyzer = GrowthAnalyzer()
+    plotter = Plotter()
+
     # read Excel file with Pandas
-    # design = analyzer.read_design(os.path.join(ROOT, args.file))
+    design = reader.read_design()
 
     # first, check if the files from the file and your computer are the same
     des_files = design['File'].to_list()
-    files_in_system = file_parser(path=ROOT,
-                                pattern='*.txt')
+    # files_in_system = file_parser(path=ROOT, pattern='*.txt') # This function is not defined in the new structure
+    files_in_system = [f for f in os.listdir(ROOT) if f.endswith('.txt')]
+
 
     print(f'Len of files in system: {len(files_in_system)}, and len of files in design: {len(des_files)}')
 
-    if check_if_equal(des_files, files_in_system):
+    # if set(des_files).issubset(set(files_in_system)): # This function is not defined in the new structure
+    if set(des_files).issubset(set(files_in_system)):
         pass
     else:
         raise Exception("The files in the folder are not the same as "
@@ -233,7 +153,7 @@ def main():
             os.makedirs(os.path.join(ROOT, OUTPUT, 'Plots'))
         else:
             print('Plots will be saved in the existing folder.')
-            
+
     # LOOP OVER THE FILES
     print(f'{bcolors.OKCYAN}Starting the analysis of the files: {bcolors.ENDC}\n')
     # parallel loop to get the AUCs
@@ -241,10 +161,10 @@ def main():
     with get_context(start_method).Pool(n_threads) as p:
         # user starmap to pass multiple arguments to the function
         out_auc_df = pd.concat(list(tqdm(p.starmap(
-                                                analyzer.calculate_auc, 
-                                                zip(design.File.to_list(), 
-                                                # modify this list to pass different arguments to the function
-                                                ['AUC']*len(design.File.to_list()))), 
+                                                analyzer.calculate_auc,
+                                                zip(design.File.to_list(),
+                                                [ROOT]*len(design.File.to_list()),
+                                                ['AUC']*len(design.File.to_list()))),
                                     total=len(design.File.to_list()))), axis=0)
     p.close()
     print('\n')
@@ -253,10 +173,9 @@ def main():
     with get_context(start_method).Pool(n_threads) as p:
         # user starmap to pass multiple arguments to the function
         out_time_df = pd.concat(list(tqdm(p.starmap(
-                                                analyzer.calculate_auc, 
-                                                zip(design.File.to_list(), 
-                                                # modify this list to pass different arguments to the function
-                                                ['timeseries']*len(design.File.to_list()))), 
+                                                processor.smooth_data,
+                                                zip(design.File.to_list(),
+                                                [ROOT]*len(design.File.to_list()))),
                                     total=len(design.File.to_list()))), axis=0)
     p.close()
     print('\n')
@@ -266,12 +185,13 @@ def main():
         patterns = design['Pattern'].unique().tolist()
         final_pattern_df = pd.DataFrame()
         for pattern in patterns:
-            pattern_vars = get_sheet_names(os.path.join(ROOT, pattern))
+            # pattern_vars = get_sheet_names(os.path.join(ROOT, pattern)) # This function is not defined in the new structure
+            pattern_vars = pd.ExcelFile(os.path.join(ROOT, pattern)).sheet_names
             pattern_df = pd.DataFrame()
             for pat in pattern_vars:
                 pattern_media = pd.read_excel(os.path.join(ROOT, pattern), pat)
                 pattern_media = pattern_media.set_index(pattern_media.columns[0])
-                well = analyzer.get_well_names(pattern_media.shape[0], pattern_media.shape[1])
+                well = processor.get_well_names(pattern_media.shape[0], pattern_media.shape[1])
                 pattern_media_vec = pattern_media.values.flatten().tolist()
                 pattern_media_df = pd.DataFrame({'Well': well, f'{pat}': pattern_media_vec, 'Pattern': pattern})
                 pattern_df = pd.concat([pattern_df, pattern_media_df], axis=1)
@@ -280,15 +200,13 @@ def main():
             pattern_df = pattern_df[['Pattern', 'Well'] + [col for col in pattern_df.columns if col not in ['Pattern', 'Well']]]
             final_pattern_df = pd.concat([final_pattern_df, pattern_df], axis=0)
 
-
-    # merge the out_auc_df with the design 
+    # merge the out_auc_df with the design
     # if 'Pattern', merge design with final_pattern_df and then merge with out_auc_df
     if 'Pattern' in design.columns:
         design_ext = design.merge(final_pattern_df, on=['Pattern'], how='right')
         out_auc_df = design_ext.merge(out_auc_df, on=['File', 'Well'], how='left')
     else:
-        out_auc_df = design.merge(out_auc_df, on='File', how='left') 
-    # out_auc_df = design.merge(out_auc_df, on='File', how='left') # old way
+        out_auc_df = design.merge(out_auc_df, on='File', how='left')
     # merge the out_time_df with the design
     # if 'Pattern', merge design with final_pattern_df and then merge with out_time_df
     if 'Pattern' in design.columns:
@@ -302,7 +220,6 @@ def main():
     # save the timeseries file in the output folder as a csv file
     out_time_df.to_csv(os.path.join(ROOT, OUTPUT, 'Timeseries.csv'), index=False)
 
-
     # PARAMETRIC MODELLING
     if args.parametric:
         print(f'\n{bcolors.OKCYAN}Starting the parametric modelling...{bcolors.ENDC}\n')
@@ -314,11 +231,11 @@ def main():
         else:
             print('The Parametric folder already exists. I will overwrite the files.')
 
-        models = {'logistic': growth, 'gompertz': gompertz}
+        models = {'logistic': analyzer.growth, 'gompertz': analyzer.gompertz}
 
         # get the timeseries data for the smoothed data
-        time_h = [int(i) for i in out_time_df.columns if is_number(i)]
-        time_h = np.array(time_h)/60/60
+        time_h = [int(i) for i in out_time_df.columns if plotter.is_number(i)]
+        time_h = [t/3600 for t in time_h]
 
         # get the smoothed data
         w_filt_df = out_time_df[out_time_df.Data.str.endswith('_f')]
@@ -330,7 +247,7 @@ def main():
             y = row[w_filt_df.columns[4:]].values
             x = time_h
 
-            best_model, best_params, best_aic = model_selection(x, y, models)
+            best_model, best_params, best_aic = analyzer.model_selection(x, y, models)
 
             if best_params is not None:
                 # create a dictionary with the results
@@ -345,13 +262,9 @@ def main():
                 parametric_df = parametric_df.append(results, ignore_index=True)
 
                 # plot the results
-                plt.figure()
-                plt.plot(x, y, 'o', label='data')
-                plt.plot(x, models[best_model](x, *best_params), '-', label='fit')
-                plt.legend()
-                plt.title(f'{row.File}_{row.Well}')
-                plt.savefig(os.path.join(parametric_path, f'{row.File}_{row.Well}.pdf'))
-                plt.close()
+                plotter.plot_parametric_fit(x, y, models[best_model], best_params,
+                                            f'{row.File}_{row.Well}',
+                                            os.path.join(parametric_path, f'{row.File}_{row.Well}.pdf'))
 
         # merge the parametric_df with the out_auc_df
         out_auc_df = out_auc_df.merge(parametric_df, on=['File', 'Well'], how='left')
@@ -371,27 +284,10 @@ def main():
         # get the smoothed data
         w_filt_df = out_time_df[out_time_df.Data.str.endswith('_f')]
 
-        # get the data to be used in the PCA
-        pca_data = w_filt_df[w_filt_df.columns[4:]].values
-
-        # perform PCA
-        pca = PCA(n_components=2)
-        pca_result = pca.fit_transform(pca_data)
-
-        # create a dataframe with the PCA results
-        pca_df = pd.DataFrame(data=pca_result, columns=['PC1', 'PC2'])
-
-        # merge with the metadata
-        pca_df = pd.concat([w_filt_df.reset_index(drop=True), pca_df], axis=1)
+        pca_df, explained_variance = analyzer.run_pca(w_filt_df, args.pca_color)
 
         # plot the results
-        plt.figure(figsize=(10, 10))
-        sns.scatterplot(x='PC1', y='PC2', data=pca_df, hue=args.pca_color, palette='viridis')
-        plt.title('PCA of growth curves')
-        plt.xlabel(f'PC1 ({pca.explained_variance_ratio_[0]:.2f})')
-        plt.ylabel(f'PC2 ({pca.explained_variance_ratio_[1]:.2f})')
-        plt.savefig(os.path.join(pca_path, 'PCA_plot.pdf'))
-        plt.close()
+        plotter.plot_pca(pca_df, args.pca_color, explained_variance, pca_path)
 
 
     ### PLOT THE TIMESERIES
@@ -402,32 +298,33 @@ def main():
 
     print(f'\nPlotting the {bcolors.OKGREEN}timeseries{bcolors.ENDC} in {print_out}. \n')
 
-    plotly_inputs = zip([out_time_df]*len(list(product(plates, data_types))), 
-                        [plate for plate in plates for i in range(len(data_types))], 
+    plotly_inputs = zip([out_time_df]*len(list(product(plates, data_types))),
+                        [plate for plate in plates for i in range(len(data_types))],
                         [data_type for i in range(len(plates)) for data_type in data_types],
                         [out_path]*len(list(product(plates, data_types))))
 
     # print(out_path)
     # loop over the plates and data types using plotly_wrapper function
     with get_context(start_method).Pool(n_threads) as p:
-        p.starmap(plotly_wrapper, tqdm(plotly_inputs, total=len(list(product(plates, data_types)))))
+        p.starmap(plotter.plotly_wrapper, tqdm(plotly_inputs, total=len(list(product(plates, data_types)))))
 
     # EXTENDED ANALYSIS
-    desing_sheets = get_sheet_names(os.path.join(ROOT, 'Design.xlsx'))
+    # desing_sheets = get_sheet_names(os.path.join(ROOT, 'Design.xlsx')) # This function is not defined in the new structure
+    desing_sheets = pd.ExcelFile(os.path.join(ROOT, 'Design.xlsx')).sheet_names
     if 'analysis' in desing_sheets:
         print('\nI found the analysis sheet in the design file. I will run the extended analysis.\n')
         analysis_vars = pd.read_excel(os.path.join(ROOT, 'Design.xlsx'), 'analysis')
-        
-        analysis_vars = AnalysisVars(analysis_vars)
-        grp_var = analysis_vars.grouping_variable
-        condition = analysis_vars.condition
+
+        # analysis_vars = AnalysisVars(analysis_vars) # This class is not defined in the new structure
+        grp_var = analysis_vars.grouping_variable.dropna().to_list()
+        condition = analysis_vars.condition.dropna().to_list()[0]
 
         # check that grp_var and condition are columns in the out_auc_df
-        if grp_var in out_auc_df.columns and condition in out_auc_df.columns:
+        if all(item in out_auc_df.columns for item in grp_var) and condition in out_auc_df.columns:
             pass
         else:
             print(f'{bcolors.FAIL}The grouping variable and/or the condition are not in the dataset. Please check the design file.{bcolors.ENDC}')
-            sys.exit()
+            # sys.exit() # This is not available in the new structure
 
         # create a folder within Plots named "Boxplots"
         boxplot_path = os.path.join(ROOT, OUTPUT, 'Plots', 'Boxplots')
@@ -438,25 +335,23 @@ def main():
 
         # clean the dataset
         out_auc_df_boxplot = out_auc_df.copy()
-        out_auc_df_boxplot.dropna(subset=[grp_var, condition], inplace=True)
+        out_auc_df_boxplot.dropna(subset=grp_var + [condition], inplace=True)
         # Vars to plot
-        temp_vars = out_auc_df_boxplot[grp_var].unique()
+        temp_vars = out_auc_df_boxplot[grp_var[0]].unique()
         # plot the y val as the f_AUC column
         y_var = [col for col in out_auc_df_boxplot.columns if col.endswith('f_AUC')][0]
 
         # store the inputs for the boxplot function in a zip object
         inputs = zip([out_auc_df_boxplot]*len(temp_vars),
-                        [grp_var]*len(temp_vars),
+                        [grp_var[0]]*len(temp_vars),
                         temp_vars,
                         [condition]*len(temp_vars),
                         [y_var]*len(temp_vars),
                         [os.path.join(ROOT, OUTPUT, 'Plots', 'Boxplots')+'/']*len(temp_vars))
 
         with get_context(start_method).Pool(n_threads) as p:
-            p.starmap(plot_boxplots, tqdm(inputs, total=len(temp_vars)))
+            p.starmap(plotter.plot_boxplots, tqdm(inputs, total=len(temp_vars)))
 
-        
-    # to test: python growth.py -i ./test_data/ -t 6
 
     print(f"\n{bcolors.BOLD}All analyses have been completed successfully!{bcolors.ENDC}\n")
